% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARI.R
\name{ARI}
\alias{ARI}
\title{Compute the adjusted rand index}
\usage{
ARI(est, truth)
}
\arguments{
\item{est}{A list of results returned from \code{khaplotype} or \code{kmodes}.}

\item{truth}{A numeric or character vector of true assignemts.}
}
\value{
A numberic value between 0 and 1, which indicates agreement between two partitions.
}
\description{
Compute the adjusted rand index given the estimated assignments and the true assiganments.
}
\examples{
# Estimate cluster assignments by function `khaplotype`.
res_khap <- khaplotype (K = 5, datafile = "../data/sim.fastq", n_init = 3)
true_assignments <- as.numeric(read.table("../data/assignment.txt", header = F, sep = ""))
ARI(res_khap, true_assignments)

}
