% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khaplotype.R
\name{khaplotype}
\alias{khaplotype}
\title{Clustering amplicon datasets with quality scores.}
\usage{
khaplotype(K = 3, datafile = "../data/sim.fastq")
}
\arguments{
\item{K}{Number of clusters. Default is 1.}

\item{datafile}{Path to a data file. Has to be a fastq file if want to conduct clustering on amplicon data.}

\item{n_init}{Number of initializations.}

\item{algorithm}{Algorithm to implement clustering. Default is "FASTQ_LLOYDS_EFFICIENT". See details for the options available.}

\item{seed}{Random number seed. Default is 1.}

\item{shuffle}{Incidate if shuffle the input order. Default is FALSE.}

\item{sim_info_file}{Path to the true observation assignments file to compute adjusted rand index.}
}
\value{
Returns a list of clustering results.
}
\description{
Implement three unsupervised clustering algorithms on amplicon datasets with quality scores.
}
\details{
Algorithms avaiable:

\itemize{
   \item \code{"FASTQ_LLOYDS_EFFICIENT"}: Efficient Lloyds algorithm
   \item \code{"FASTQ_HW_EFFICIENT"}: Efficient Hartigan and Wong algorithm
   \item \code{"FASTQ_MACQUEEN"}: MacQueen's algorithm
   \item \code{"FASTQ_LLOYDS"}: Lloyds algorithm
   \item \code{"FASTQ_HW"}: Hartigan and Wong algorithm
 }

Value:
\itemize{
   \item \code{"best_cluster_size"}: Number of observations in each cluster of the best initialization.
   \item \code{"best_criterion"}: Optimized criterion in each cluster of the best initialization.
   \item \code{"best_cluster_id"}: Cluster assignment of each observation of the best initialization.
   \item \code{"best_modes"}: Estimated modes for each cluster of the best initialization.
   \item \code{"total_best_criterion"}: Total optimized criterion of the best initialization.
   \item \code{"clsuter_size"}: Number of clusters.
   \item \code{"data_dim"}: Dimension of input data.
}
}
\examples{
# Clustering an amplicon dataset and run three initializations with default algorithm ("FASTQ_HW_EFFICIENT")
res_khap <- khaplotype(K = 5, datafile = "../data/sim.fastq", n_init = 3)

# Clustering an amplicon dataset and run three initializations with MacQueen's algorithm (shuffle the data)
res_khap <- khaplotype(K = 5, datafile = "../data/sim.fastq", n_init = 3,
algorithm = "FASTQ_MACQUEEN", shuffle = TRUE)

# Clustering an amplicon dataset provide a different seed
res_khap <- khaplotype(K = 5, datafile = "../data/sim.fastq", seed = 1)

}
