% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster.R
\name{plot_cluster}
\alias{plot_cluster}
\title{Plotting clusters after using \code{kmodes} or \code{khaplotype} function.}
\usage{
plot_cluster(dat, res)
}
\arguments{
\item{dat}{Input data.}

\item{res}{Results returned from \code{kmodes} or \code{khaplotype}.}
}
\value{
A plot reflecting clustering results.
}
\description{
Visulize clusters after using the cluster algorithms implemented in the functions.
This cluster plot make use of the function \code{dapc} from package \code{adegenet}
and function \code{scatter} from package \code{ade4}. The plot is based on observation
assignment and discriminant analysis of principal components, therefore, when running
the code, it will asks you to choose the number PCs to retain and choose the number
discriminant functions to retain.
}
\examples{
# use function \\code{kmodes}
data <- read.table("../data/zoo.int.data")
res_kmodes <- kmodes(K = 5, datafile = "../data/zoo.int.data",
algorithm = "KMODES_HARTIGAN_WONG", init_method = "KMODES_INIT_AV07_GREEDY",
n_init = 10)
plot_cluster(data, res_kmodes)

}
