% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_fasta}
\alias{read_fasta}
\title{Read a fasta file}
\usage{
read_fasta(datafile = NULL)
}
\arguments{
\item{datafile}{Path to a fasta file.}
}
\value{
Return a list contains reads and dim of the data.
}
\description{
Read a fasta file (Reads need to be the same length).
}
\examples{
# Read a fasta file
datFile <- system.file("extdata", "sim.fasta", package = "CClust")
dat <- read_fasta(datafile = datFile)
}
