% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_fastq}
\alias{read_fastq}
\title{Read a fastq file}
\usage{
read_fastq(datafile = NULL)
}
\arguments{
\item{datafile}{Path to a fastq file.}
}
\value{
Return a list contains reads, quality socres and dim of the data.
}
\description{
Read a fastq file and encode Phred quality score from 0 to 93 using ASCII 33 to 126.
}
\examples{
# Read a fastq file
datFile <- system.file("extdata", "sim.fastq", package = "CClust")
dat <- read_fastq(datafile = datFile)
}
