% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster.R
\name{plot_cluster}
\alias{plot_cluster}
\title{Plotting clusters after using \code{kmodes} or \code{khaplotype} function.}
\usage{
plot_cluster(res, isGene = FALSE)
}
\arguments{
\item{res}{Results returned from \code{kmodes} or \code{khaplotype}.}

\item{isGene}{Indicate if the clustering data is gene sequences, default if FALSE.}
}
\value{
A plot reflecting clustering results.
}
\description{
Visulize clusters after using the cluster algorithms implemented in the functions.
This cluster plot make use of the function \code{dapc} from package \code{adegenet}
and function \code{scatter} from package \code{ade4}. The plot is based on observation
assignment and discriminant analysis of principal components, therefore, when running
the code, it will asks you to choose the number PCs to retain and choose the number
discriminant functions to retain.
}
\examples{
# use function \\code{kmodes}
\dontrun{
data <- system.file("extdata", "zoo.int.data", package = "CClust")
res_kmodes <- kmodes(K = 5, datafile = data, algorithm = "KMODES_HARTIGAN_WONG",
init_method = "KMODES_INIT_AV07_GREEDY", n_init = 10)
plot_cluster(res_kmodes)
}

# use function \\code{khaplotype}
\dontrun{
data <- system.file("extdata", "sim_small.fastq", package = "CClust")
res_khap <- khaplotype(K = 5, datafile = data, n_init = 10)
plot_cluster(res_khap, isGene = TRUE)
}
}
