% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulator}
\alias{simulator}
\title{Simulate clustered categorical datasets.}
\usage{
simulator(simK, n_coordinates, n_observations,
n_categories, sim_between_t, sim_within_t, use_dirichlet = FALSE, sim_pi)
}
\arguments{
\item{simK}{Number of clusters.}

\item{n_coordinates}{Number of coordinate of the simulated dataset.}

\item{n_observations}{Number of observation of the simulated dataset.}

\item{n_categories}{Number of categories of the simulated dataset.}

\item{sim_between_t}{Between cluster variation.}

\item{sim_within_t}{Within cluster variation.}

\item{use_dirichlet}{Indicate if cimulate datasets with dirichlet prior. Defalut is FALSE.}

\item{sim_pi}{Mixing proportions, a vector with the same length of specified number of clusters
and the sum of the values in this vector has to be 1.}
}
\value{
Returns a list of simulation dataset results.
}
\description{
Simulate clustered categorical datasets by using continous time Marcov chain.
}
\details{
Value:
\itemize{
   \item \code{"CTMC_probabilities"}: Number of observations in each cluster of the best initialization.
   \item \code{"modes"}: Simulated modes.
   \item \code{"cluster_assignments"}: Simulated cluster assignments.
   \item \code{"cluster_sizes"}: Simulated cluster sizes.
   \item \code{"data"}: Simulated data.
}
}
\examples{
#Simulate data with dim 100 * 10, 4 different categories and there are 5 true clusters.
data <- simulator(simK = 5, n_coordinates = 10, n_observations = 100, n_categories = 4,
sim_between_t = 2, sim_within_t = 1, use_dirichlet = TRUE, sim_pi = c(0.1, 0.1, 0.2, 0.3, 0.3))
}
